feat: re-enable forge fmt + foundry v1.5.0 (#1669)
**Motivation:**

Foundry 1.4.0 contained many breaking changes for forge fmt, thus we
opted to temporarily disable it to avoid massive diffs.

**Modifications:**

- Bumped ci v1.3.5 -> v1.5.0
- Added all forge fmt config vars to foundry.toml
- Ran make fmt

**Result:**

CI checks that code is formatted with v1.5.0.

Added variables:
`style` - Indent style (space/tab)
`hex_underscore` - Underscore style in hex literals
`override_spacing` - Spacing in override attributes
`docs_style` - Documentation comment style
`contract_new_lines` - New lines at contract boundaries
`pow_no_space` - Spacing around power operator ()
`prefer_compact` - List element breaking strategy
`single_line_imports` - Keep imports on single lines

---------

+++ src/test/mocks/MockERC20.sol
@@ -224,8 +224,8 @@ contract MockERC20 {
     }
 
     function _pureChainId() private pure returns (uint chainId) {
-        function() internal view returns (uint256) fnIn = _viewChainId;
-        function() internal pure returns (uint256) pureChainId;
+        function() internal view returns (uint) fnIn = _viewChainId;
+        function() internal pure returns (uint) pureChainId;
         assembly {
             pureChainId := fnIn
         }
